PlayerListener refactored
